+++ packages/eyes-sdk-core/lib/runner/EyesRunner.js
@@ -1,6 +1,7 @@
 'use strict'
 
 const {GeneralUtils} = require('@applitools/eyes-common')
+const getScmInfo = require('../getScmInfo')
 
 class EyesRunner {
   constructor() {
@@ -8,13 +9,32 @@ class EyesRunner {
     this._eyesInstances = []
   }
 
+  makeGetBatchInfo(fetchBatchInfo) {
+    if (!this._getBatchInfo) {
+      this._getBatchInfo = GeneralUtils.cachify(fetchBatchInfo)
+    }
+  }
+
+  async getBatchInfoWithCache(batchId) {
+    if (this._getBatchInfo) {
+      return this._getBatchInfo(batchId)
+    } else {
+      throw new Error(
+        'Eyes runner could not get batch info since makeGetBatchInfo was not called before',
+      )
+    }
+  }
+
+  async getScmInfoWithCache(...args) {
+    return getScmInfo(...args)
+  }
+
   /**
    * @abstract
-   * @param {boolean} [shouldThrowException=true]
-   * @return {Promise<TestResultSummary>}
+   * @param {boolean} [throwEx=true]
+   * @return {Promise<TestResultsSummary>}
    */
-  async getAllTestResults(_shouldThrowException) {
-    // eslint-disable-line no-unused-vars
+  async getAllTestResults(_throwEx) {
     throw new TypeError('The method is not implemented!')
   }
 
@@ -35,22 +55,6 @@ class EyesRunner {
       }
 
       await Promise.all(promises)
-    }
-  }
-
-  makeGetBatchInfo(fetchBatchInfo) {
-    if (!this._getBatchInfo) {
-      this._getBatchInfo = GeneralUtils.cachify(fetchBatchInfo)
-    }
-  }
-
-  async getBatchInfoWithCache(batchId) {
-    if (this._getBatchInfo) {
-      return this._getBatchInfo(batchId)
-    } else {
-      throw new Error(
-        'Eyes runner could not get batch info since makeGetBatchInfo was not called before',
-      )
     }
   }
 }