+++ packages/eyes-webdriverio-4/src/EyesWDIO.js
@@ -21,6 +21,7 @@ const {
   ArgumentGuard,
   SimplePropertyHandler,
   Configuration,
+  ClassicRunner,
 } = require('@applitools/eyes-sdk-core')
 
 const {DomCapture} = require('@applitools/dom-utils')
@@ -43,7 +44,6 @@ const Target = require('./fluent/Target')
 const WDIOJSExecutor = require('./WDIOJSExecutor')
 const WebDriver = require('./wrappers/WebDriver')
 const ReadOnlyPropertyHandler = require('@applitools/eyes-sdk-core/index').ReadOnlyPropertyHandler
-const {ClassicRunner} = require('./runner/ClassicRunner')
 const ImageRotation = require('./positioning/ImageRotation')
 
 const VERSION = require('../package.json').version
@@ -1010,7 +1010,10 @@ class EyesWDIO extends EyesBase {
    * @return {Promise<TestResults>}
    */
   async close(throwEx = true) {
-    const results = await super.close(throwEx)
+    const results = await super.close(true).catch(err => {
+      if (!throwEx) return err
+      else throw err
+    })
 
     if (this._runner) {
       this._runner._allTestResult.push(results)